requirements.txt
customtkinter>=5.2.0,<6.0.0
yt-dlp>=2024.3.0,<2025.0.0
psutil>=5.9.0,<6.0.0
requests>=2.31.0,<3.0.0
Pillow>=10.0.0
